+++ packages/@groupby/elements-cache-driver-plugin/test/unit/common/index.test.ts
@@ -1,0 +1,9 @@
+import { expect } from '../../utils';
+import { CacheDriverPlugin as CacheDriverPluginExport } from '../../../src';
+import CacheDriverPlugin from '../../../src/cache-driver-plugin';
+
+describe('Entry point', () => {
+  it('should export the CacheDriverPlugin', () => {
+    expect(CacheDriverPluginExport).to.equal(CacheDriverPlugin);
+  });
+});