+++ packages/@groupby/elements-cache-plugin/test/unit/common/index.test.ts
@@ -1,0 +1,9 @@
+import { expect } from '../../utils';
+import { CachePlugin as CachePluginExport } from '../../../src';
+import CachePlugin from '../../../src/cache-plugin';
+
+describe('Entry point', () => {
+  it('should export the CachePlugin', () => {
+    expect(CachePluginExport).to.equal(CachePlugin);
+  });
+});